Separate task in different classes
Each task in a separated file
