+++ build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 }
 
 group = "com.klaxit.hiddensecrets"
-version = "0.1.1"
+version = "0.1.2"
 
 repositories {
     mavenCentral()
