+++ src/main/kotlin/com/klaxit/hiddensecrets/CodeGenerator.kt
@@ -1,15 +1,14 @@
 package com.klaxit.hiddensecrets
 
 /**
- * Helper to generate C++ and Kotlin code
+ * Helper to generate C++ and Kotlin code.
  */
 object CodeGenerator {
 
     /**
-     * Return the C++ code to add another key to your project
+     * Return the C++ code to add another key to your Android project.
      */
     fun getCppCode(packageName: String, keyName: String, obfuscatedKey: String): String {
-
         return "\nextern \"C\"\n" +
                 "JNIEXPORT jstring JNICALL\n" +
                 "Java_" + Utils.getSnakeCasePackageName(packageName) + "_Secrets_get$keyName(\n" +
@@ -22,7 +21,7 @@ object CodeGenerator {
     }
 
     /**
-     * Kotlin code that will be added in your project
+     * Kotlin code that will be added in your Android project.
      */
     fun getKotlinCode(keyName: String): String {
         return "\n    external fun get$keyName(packageName: String): String\n" +
