+++ src/main/kotlin/com/klaxit/hiddensecrets/Constants.kt
@@ -1,0 +1,9 @@
+package com.klaxit.hiddensecrets
+
+/**
+ * Constants used several time in the plugin.
+ */
+object Constants {
+    const val APP_MAIN_FOLDER = "src/main/"
+    const val KOTLIN_FILE_NAME = "Secrets.kt"
+}
