+++ src/main/kotlin/com/klaxit/hiddensecrets/task/ObfuscateTask.kt
@@ -1,0 +1,14 @@
+package com.klaxit.hiddensecrets.task
+
+import org.gradle.api.tasks.TaskAction
+
+/**
+ * Obfuscate a key, does not add it to your project. Useful to debug.
+ */
+open class ObfuscateTask: BaseTask() {
+
+    @TaskAction
+    fun obfuscate() {
+        getObfuscatedKey()
+    }
+}
